[ETFE-1499] moved TraderModel and AddressModel from reportOfReceipt to common
