+++ app/uk/gov/hmrc/emcstfe/models/request/SubmitReportOfReceiptRequest.scala
@@ -18,7 +18,8 @@ package uk.gov.hmrc.emcstfe.models.request
 
 import uk.gov.hmrc.emcstfe.models.auth.UserRequest
 import uk.gov.hmrc.emcstfe.models.common.DestinationType.{DirectDelivery, RegisteredConsignee, TaxWarehouse, TemporaryRegisteredConsignee}
-import uk.gov.hmrc.emcstfe.models.reportOfReceipt.{SubmitReportOfReceiptModel, TraderModel}
+import uk.gov.hmrc.emcstfe.models.common.TraderModel
+import uk.gov.hmrc.emcstfe.models.reportOfReceipt.SubmitReportOfReceiptModel
 
 import java.time.{LocalDate, LocalTime, ZoneId}
 import java.util.UUID
