+++ test/uk/gov/hmrc/emcstfe/fixtures/GetMovementFixture.scala
@@ -18,7 +18,8 @@ package uk.gov.hmrc.emcstfe.fixtures
 
 import play.api.libs.json.{JsValue, Json}
 import uk.gov.hmrc.emcstfe.models.common.DestinationType.Export
-import uk.gov.hmrc.emcstfe.models.reportOfReceipt.{AddressModel, ConsignorTraderModel, TraderModel}
+import uk.gov.hmrc.emcstfe.models.common.{AddressModel, TraderModel}
+import uk.gov.hmrc.emcstfe.models.reportOfReceipt.ConsignorTraderModel
 import uk.gov.hmrc.emcstfe.models.response.{GetMovementResponse, MovementItem, Packaging, WineProduct}
 
 trait GetMovementFixture extends BaseFixtures {
@@ -469,6 +470,7 @@ trait GetMovementFixture extends BaseFixtures {
     sequenceNumber = 1,
     destinationType = Export,
     consigneeTrader = Some(TraderModel(
+      vatNumber = None,
       traderId = Some("GB11100000002"),
       traderName = Some("Current 801 Consignee"),
       address = Some(AddressModel(
