+++ test/uk/gov/hmrc/emcstfe/fixtures/GetMovementIfChangedFixture.scala
@@ -17,7 +17,8 @@
 package uk.gov.hmrc.emcstfe.fixtures
 
 import uk.gov.hmrc.emcstfe.models.common.DestinationType.Export
-import uk.gov.hmrc.emcstfe.models.reportOfReceipt.{AddressModel, ConsignorTraderModel, TraderModel}
+import uk.gov.hmrc.emcstfe.models.common.{AddressModel, TraderModel}
+import uk.gov.hmrc.emcstfe.models.reportOfReceipt.ConsignorTraderModel
 import uk.gov.hmrc.emcstfe.models.response.{GetMovementResponse, MovementItem, Packaging, WineProduct}
 
 trait GetMovementIfChangedFixture extends BaseFixtures {
@@ -493,6 +494,7 @@ trait GetMovementIfChangedFixture extends BaseFixtures {
     sequenceNumber = 1,
     destinationType = Export,
     consigneeTrader = Some(TraderModel(
+      vatNumber = None,
       traderId = Some("GB11100000002"),
       traderName = Some("Current 801 Consignee"),
       address = Some(AddressModel(
@@ -504,6 +506,7 @@ trait GetMovementIfChangedFixture extends BaseFixtures {
       eoriNumber = None
     )),
     deliveryPlaceTrader = Some(TraderModel(
+      vatNumber = None,
       traderId = Some("GB11100000003"),
       traderName = Some("Delivery Place"),
       address = Some(AddressModel(
