+++ test/uk/gov/hmrc/emcstfe/fixtures/TraderModelFixtures.scala
@@ -16,36 +16,38 @@
 
 package uk.gov.hmrc.emcstfe.fixtures
 
-import uk.gov.hmrc.emcstfe.models.reportOfReceipt.TraderModel
-
+import uk.gov.hmrc.emcstfe.models.common.TraderModel
 import scala.xml.NodeSeq
 
 trait TraderModelFixtures extends BaseFixtures with AddressModelFixtures {
 
   val traderId = "GB0000000012346"
 
-  val maxTraderModel = TraderModel(
+  val maxTraderModel: TraderModel = TraderModel(
+    vatNumber = Some("number"),
     traderId = Some(traderId),
     traderName = Some("name"),
     address = Some(maxAddressModel),
     eoriNumber = Some("eori")
   )
 
-  val maxTraderModelXML =
+  val maxTraderModelXML: NodeSeq =
     NodeSeq.fromSeq(Seq(
+      Seq(<urn:VatNumber>number</urn:VatNumber>),
       Seq(<urn:Traderid>{traderId}</urn:Traderid>),
       Seq(<urn:TraderName>name</urn:TraderName>),
       maxAddressModelXML,
       Seq(<urn:EoriNumber>eori</urn:EoriNumber>)
     ).flatten)
 
-  val minTraderModel = TraderModel(
+  val minTraderModel: TraderModel = TraderModel(
+    vatNumber = None,
     traderId = None,
     traderName = None,
     address = None,
     eoriNumber = None
   )
 
-  val minTraderModelXML = NodeSeq.Empty
+  val minTraderModelXML: NodeSeq = NodeSeq.Empty
 
 }
